main_code/requirements.txt
dill==0.3.4
multiprocess==0.70.12.2
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pathos==0.2.8
Pillow==9.0.1
pox==0.3.0
ppft==1.6.6.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqldiff==0.0.4
sqlparse==0.4.1
